feat: refactored custom config handling
