+++ lib/validateConfig.js
@@ -90,7 +90,7 @@ export const validateConfigLogic = (config, configPath, logger) => {
       errors.push(
         configurationError(
           pattern,
-          'If the task is an object, it should contain `title` and `task` fields, where `title` should be a string and `task` should be a function.',
+          'Custom task should contain `title` and `task` fields, where `title` should be a string and `task` should be a function.',
           task
         )
       )
